feat: allow adding screenshots without baseline

+++ src/local-server/controller.ts
@@ -60,6 +60,11 @@ export class Controller {
   }
 
   private async updateScreenshots(test: Test) {
+    // allow for empty baseline
+    if (!test.baselinePath) {
+      test.baselinePath = test.comparisonPath.replace('/comparison/', '/baseline/')
+    }
+
     // copy comparison to baseline
     try {
       await fs.copyFile(
@@ -72,7 +77,9 @@ export class Controller {
 
     // remove diff screenshot
     try {
-      await fs.remove(path.join(process.cwd(), test.diffPath))
+      if (test.diffPath) {
+        await fs.remove(path.join(process.cwd(), test.diffPath))
+      }
     } catch {
       throw Error('Cannot remove diff screenshot')
     }
